+++ handlers/images.js
@@ -34,7 +34,7 @@ function handleImage(message, postedTime, wasDelayed){
 					} else res();
 				});
 				writeFile.then(() => {
-					const logs = (ops.logsChannel != "0") ? message.client.channels.cache.get(ops.logsChannel) : undefined;
+					const logs = (ops.logsChannel) ? message.client.channels.cache.get(ops.logsChannel) : undefined;
 					if (ops.performanceMode) performanceLogger(`#${imgStats.imageLogCount + 1}: Crop started\t`, postedTime.getTime());
 					crop(message.attachments.first()).then((imgBuff) => {
 						if (ops.performanceMode) performanceLogger(`#${imgStats.imageLogCount + 1}: Crop finished\t`, postedTime.getTime());